Rename files to "counter*" from "numbers*"
