+++ src/lib/counter.ts
